pyproject.toml
[project]
authors = []
name = "MCPMark"
requires-python = ">= 3.11"
version = "0.0.1"
dependencies = [
  "notion-client>=2.4.0",
  "playwright>=1.43.0",
  "seaborn>=0.12.0",
  "matplotlib>=3.7.0",
  "numpy>=1.23.0",
  "openai-agents>=0.2.3,<0.3",
  "openai==1.96.1",
  "python-dotenv>=1.1.1,<2",
  "ruff>=0.12.4,<0.13",
  "psycopg2-binary>=2.9.10,<3",
  "pyyaml>=6.0.2,<7"
, "nest-asyncio>=1.6.0,<2", "pixi", "pipx>=1.7.1,<2", "pgdumplib>=3.1.0,<4"]

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = [
  "osx-arm64",
  "linux-aarch64",
  "linux-64",
  "win-64",
  "osx-64",
]

[tool.pixi.tasks]
fmt = "ruff"

[tool.ruff.format]
indent-style = "space"
line-ending = "auto"

[tool.hatch.build.targets.wheel]
packages = ["src", "tasks"]
